ccadb-webpki-roots: add webpki root generator.
This commit introduces a new ccadb tool: `ccadb-webpki-roots`. It can be
run with `cargo gen-roots` after having run `cargo
fetch-included-roots`.

This tool processes the `IncludedCACertificateReportPEMCSV` report from
CCADB, filtering out root certificates that are:

* Included in the Mozilla root program.
* Trusted for websites (TLS).
* Not distrusted for TLS based on the current date.

The filtered set of trusted TLS root certificates are processed and
rendered to a `lib.rs` file matching the format expected by the webpki
crate. This file can be used as the contents of the `webpki-roots`
crate.